Rag_Project/requirements.txt
streamlit
ollama
langchain
langchain-community
chromadb
watchdog
unstructured[md]
pypdf
tiktoken
psutil
rich
python-dotenv
